refactor(fe): extract adopt phases and errors libs
